% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brassidata.R
\docType{data}
\name{Brassidata}
\alias{Brassidata}
\title{Single gene trees, a reference tree and a table of grouping of Brassicaceae.}
\format{A list of three, containing test trees, grouping table, and a ref tree:
\describe{
  \item{trees}{A list of 30 objects of class "phylo". Single gene trees.}
  \item{taxa}{A dataset containing the names of tips and their groupings.}
  \item{ref}{An object of class "phylo". A reference tree resulted from
      coalescent method.}
}}
\usage{
Brassidata
}
\description{
A list of three objects for \code{youshu}.
}
\keyword{datasets}
