% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeFilter.R
\name{cladeFilter}
\alias{cladeFilter}
\title{Filtering Sequences of Paralogy or With Other Misleading Signals Based on Given Groupings}
\usage{
cladeFilter(trees, taxa, level = NULL, write.table = TRUE)
}
\arguments{
\item{trees}{a list of objects of class "\code{phylo}".
A list of test trees such as single gene trees.}

\item{taxa}{a \code{data.frame} with tips and groupings. The first column must be
the names of all tips in \code{trees}. Other columns specify the groupings.}

\item{level}{a character specifying the level used as grouping. Should be one of the
column names of \code{taxa}. Defaults as \code{NULL} and use the second column
of \code{taxa}.}

\item{write.table}{a logical specifying whether to write out text files
containing the number and names of sequences suggested to be removed
by \code{cladeFilter} for each grouping.}
}
\value{
A list of cladeFilter result for all test trees. For a test tree, the result constains:
result    result of cladeFilter
removingTips    tip(s) violate the grouping
remainingTips    tip(s) group
mrcaInTree    tips in the subclade of the MRCA node of the grouping in the test tree
node    MRCA node for the grouping
subtree    the extracted subclade of the grouping
}
\description{
This function judges whether or not members of a grouping defined by \code{taxa}
and \code{level} grouped in \code{trees}, and suggest sequences to be removed when they are not.
Rationales for the selection of excluding sequences see \code{Details}.
}
\details{
To determine a scenario as the best solution of which sequences to be removed
    when members of a grouping is separated into more than one sub-clade, \code{cladeFilter}
    first calculate for each subclade the number of tips of a grouping to be remained or
    removed in the tree, and choose the one with the highest benefit as [number of
    remaining tips minus that of removing tips] (\code{max.benefit}). If this number is the
    same in more than one subclade, then \code{cladeFilter} chooses the one removing the
    least number of tips (\code{min.rmtips}). If both \code{max.benefit} and \code{min.rmtips}
    in more than one subclade are the same, \code{cladeFilter} searches the lengths of branches
    subtending to these subclades, remains the one with shortest branch length and
    removes the others. If the branch lengths are still the same for more than one
    subclade, \code{cladeFilter} then suggests to remove all tips of this grouping
    because the possibilities for these subclades to be paralogs are similar.
}
\examples{
\dontrun{
data(Brassidata)
trees <- Brassidata$trees
taxa <- Brassidata$taxaTable
# perform cladeFilter analysis
ures <- cladeFilter(trees = trees, taxa = taxa, level = 2, write.table = FALSE)
# summarize cladeFilter results with multiple groupings
usum <- summarizeMulticf(ures, trees, plot.tree = FALSE, write.table = FALSE)
# to make barplots
utest <- barplotMulticf(x = usum, reorder="up", plot = TRUE, pdf=FALSE)

}
}
\seealso{
\code{\link{summarizeMulticf}}, \code{\link{barplotMulticf}}, \code{\link{plot.cladeFilter}}
}
