% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concor.node.R
\name{concor.node}
\alias{concor.node}
\title{Concordances of Trees with All Nodes of A Reference Tree}
\usage{
concor.node(ref, trees, bp = 0, getTreeNames = FALSE, node = NULL)
}
\arguments{
\item{ref}{an object of class "\code{phylo}". A reference tree.}

\item{trees}{a list of objects of class "\code{phylo}".
A list of test trees such as single gene trees.}

\item{bp}{a bootstrape value, which the concordance is counted
when the corresponding node label of \code{trees} is higher than.}

\item{getTreeNames}{a logical specifying whether to return the names of \code{trees}
showing concordances for each node on \code{ref}. Defaults to \code{FALSE}.}

\item{node}{a numeric or integer specifying the node(s) on \code{ref}
to analyze the concordance. Defaults to \code{NULL} and all the nodes on
\code{ref} are analyzed.}
}
\description{
\code{concor.node} analyzes and returns for each node on the ref tree
the number and names of \code{trees} showing concordances.
}
\details{
Concordance for a node is defined as all descendent tips of the node of \code{ref}
    are grouped together without other tips in \code{trees}. The main analysis of this
    function requires rooted trees; when the \code{trees} are not rooted, \code{concor.node}
    roots \code{trees} according to \code{ref} during the analysis. \code{concor.node}
    returns an object of class "\code{concor}" as a list of: (1) \code{tree}: the \code{ref};
    (2) data: a \code{data.frame} showing the counts and percents of concordances for each
    node; (3) \code{conTrees}: names of \code{trees} that are concordant with \code{ref} for
    each node (when \code{getTreeNames = TRUE}).
}
\examples{
\dontrun{
data(Brassidata)
trees <- Brassidata$trees
ref <- Brassidata$ref
concor.n<-concor.node(ref = ref,trees = trees, bp = c(0,30,50), getTreeNames = FALSE)
}
}
