% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concor.trees.R
\name{concor.trees}
\alias{concor.trees}
\title{Analyze Concordances on Trees with All Groupings}
\usage{
concor.trees(trees, taxa, level = NULL)
}
\arguments{
\item{trees}{a list of objects of class "\code{phylo}".
A list of test trees such as single gene trees.}

\item{taxa}{a \code{data.frame} with tips and groupings. The first column must be
the names of all tips in \code{trees}. Other columns specify the groupings.}

\item{level}{a character specifying the level used as grouping. Should be one of the
column names of \code{taxa}. Defaults as \code{NULL} and use the second column
of \code{taxa}.}
}
\description{
An internal function of \code{backboneBP}. To summarize the concordances
with all groupings of each of \code{trees}, and returns trees that meets all groupings
to the main analysis of \code{backboneBP}.
}
\examples{
\dontrun{
data(Brassidata)
trees <- Brassidata$trees
taxa <- Brassidata$taxaTable
uconcort <- concor.trees(trees, taxa, level = 2)
}
}
\seealso{
\code{\link{backboneBP}}
}
