% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGrouping.R
\name{createGrouping}
\alias{createGrouping}
\title{Creating a Taxa Table as Input to Other Functions in package \code{youshu}}
\usage{
createGrouping(concor, percent = 0.5, count = NULL, bp = NULL, nodes = NULL)
}
\arguments{
\item{concor}{an object from "\code{concor.node}".}

\item{percent}{a numeric between 0 to 1. The groupings are created from nodes
showing the percent of concordances above \code{percent}. Defaults to \code{0.5}.}

\item{count}{a numeric or integer higher than or equals to 1.
The groupings are created from nodes showing the count of concordances
above \code{count}. Defaults to \code{NULL} and uses \code{percent}.}

\item{bp}{a numeric specifying with which bp the result in \code{concor} is used
to create \code{taxa}. Defaults to \code{NULL} and the first result in
\code{concor} is used.}

\item{nodes}{a numeric vector specifying the node(s) to create \code{taxa},
regardless of \code{percent} and \code{count}. Defaults to \code{NULL}.}
}
\description{
By giving \code{percent} or \code{count}, the groupings are created from nodes
    showing the percent or count of concordances above \code{percent} or \code{count},
    respectively, and returns a \code{data.frame} that can be used directly as \code{taxa}
    in functions \code{cladeFilter} and \code{backboneBP}. Also can specify the nodes to
    produce groupings by giving \code{nodes} directly.
}
\examples{
\dontrun{
data(Brassidata)
trees <- Brassidata$trees
ref <- Brassidata$ref
concor.n<-concor.node(ref = ref,trees = trees, bp = c(0,30,50), getTreeNames = FALSE)
# to get groupings from nodes with >70\% tree in concordance
taxa <- createGrouping(concor.n, percent = 0.7)
}
}
\seealso{
\code{\link{concor.node}}
}
