% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUnref.R
\name{createUnref}
\alias{createUnref}
\title{createUnref}
\usage{
createUnref(concor, bp, percent)
}
\arguments{
\item{concor}{an object of class "\code{concor.node}".}

\item{bp}{a numeric specifying the bp value of node(s) to be set as unresolved node
when the corresponding node labels are lower than \code{bp}.}

\item{percent}{a numeric between 0 to 1. The constraints are created from nodes
showing the percent of concordances above \code{percent}. Defaults to \code{0.5}.}
}
\description{
Internal function of \code{BetterTree}.
}
