% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_closest_out_for_GD.R
\name{detect_closest_out_for_GD}
\alias{detect_closest_out_for_GD}
\title{Detect the existence of the closest outgroups for each gd}
\usage{
detect_closest_out_for_GD(
  phyto_node,
  trees,
  gdtable,
  sptree,
  split,
  tree_id_tab,
  up_one_node = FALSE,
  outgroup_for_tree = NULL,
  pdfwid = 10,
  pdflen = 30
)
}
\arguments{
\item{phyto_node}{a table with the first column as the phyto id given by Tree2GD and 
the second column as the node id from \code{ape}.}

\item{trees}{a list of objects of class "\code{phylo}".
A list of gene trees to be examined.}

\item{gdtable}{a \code{data.frame} of the gd table file provided by Tree2GD.}

\item{sptree}{the species tree used to do tree reconciliation.}

\item{split}{a character. The symbol used to separate species name and the 
sequence number in the gene family trees.}

\item{tree_id_tab}{a table for the tree id during Tree2GD and the name of the 
corresponding tree file. The first column is the id, and the second column
is the tree names.}

\item{up_one_node}{logical. Whether to choose tips of one node deeper as the 
closest outgroup, as a second step of investigation. Default \code{FALSE}.}

\item{outgroup_for_tree}{optional. The outgroup for the whole species tree.
If given, these outgroups will be colored as yellow.}

\item{pdfwid}{the width of pdf file for subtrees. Default \code{10}.}

\item{pdflen}{the length of pdf file for subtrees. Default \code{30}.}
}
\description{
\code{detect_closest_out_for_GD} is to find out whether each gd 
    has tips of its sister clade as the closest outgroup, in order to verify 
    the support of gd. The returning object contains a simple summary for the 
    number and percent of gd that has the closest outgroup. This function will 
    also produce pdf files plotting subclade of each gd, with tips colored, 
    for easier manual investigation when required.
}
\details{
For each gd, detect whether there are closest outgroups
   exactly next to the duplicated node, that can actually define 
   the gd at the indicated node. Otherwise, gd will be possible 
   to be at other (depper) nodes as well.
}
\examples{
\dontrun{
res<-
detect_closest_out_of_GD(
phyto_node,trees = trees,gdtable = gdtable,
sptree = sptree,split=".",tree_id_tab = tree_id_tab,
up_one_node=FALSE,outgroup_for_tree=out[,1],pdfwid = 20,pdflen = 100)

res
}
}
