% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concor.trees.R
\name{.check.concor}
\alias{.check.concor}
\title{.check.concor}
\usage{
.check.concor(tre, taxa, level)
}
\arguments{
\item{tre}{tre}

\item{taxa}{taxa}

\item{level}{level}
}
\description{
Internal function of \code{youshu}
}
