% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeFilter.R
\name{.cladefilter}
\alias{.cladefilter}
\title{.cladeFilter}
\usage{
.cladefilter(trees, xx, warn, fns)
}
\arguments{
\item{trees}{trees}

\item{xx}{xx}

\item{warn}{warn}

\item{fns}{fns}
}
\description{
Internal function of \code{BetterTree}
}
