% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concor.node.R
\name{.concor.node.unrooted}
\alias{.concor.node.unrooted}
\title{.concor.node.unrooted}
\usage{
.concor.node.unrooted(ref, trees, bp = 0, getTreeNames = FALSE,
  node = NULL, warn = TRUE, unresolve.ref = FALSE,
  unresolve.para = c(bp = NA, percent = 0.8), from)
}
\arguments{
\item{ref}{ref}

\item{trees}{trees}

\item{bp}{bp}

\item{getTreeNames}{getTreeNames}

\item{node}{node}

\item{warn}{warn}

\item{unresolve.ref}{unresolve.ref}

\item{unresolve.para}{unresolve.para}

\item{from}{from}
}
\description{
Internal function of \code{youshu}
}
