% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominator.R
\name{.deno}
\alias{.deno}
\title{internal function}
\usage{
.deno(xx, ref, testtree, testtips, c.out, split, sub_coverage, max.tip, bp)
}
\arguments{
\item{xx}{one of allnodes}

\item{ref}{ref}

\item{testtree}{testtree}

\item{testtips}{testtips}

\item{c.out}{logical. Whether the closest outgroup species
are required to exist. Defaults to \code{FALSE}.}

\item{split}{a character indicates the symbol used to separate
species names and gene id for gene family trees. 
Defaults to \code{NULL}.}

\item{sub_coverage}{a numeric to specify the species coverage 
of the clades of each node. Used as percent when <= 1, 
and as count when > 1. Defaults to \code{NULL}.}

\item{max.tip}{an integer specifying the max number of tips
for the coverage filtering. Applied only when sub_coverage 
is set as percent (<= 1). Used as a cutoff for coverage as 
N tips as for deep nodes having large clades.
Defaults to \code{NULL}.}

\item{bp}{a numeric specify the minimum bp value for the node.
Defaults to \code{NULL}.}
}
\description{
Internal function of \code{rePhylo}
}
\seealso{
\code{\link{concor.node}} 
    \code{\link{detect_closest_out_for_GD}}
    \code{\link{filter_GD_by_coverage}}
}
