% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_GD_by_coverage.R
\name{.filter_coverage}
\alias{.filter_coverage}
\title{Internal. Filters the GD counts by the species coverage of
    each of the duplicated subclades}
\usage{
.filter_coverage(
  x,
  phyto_node,
  trees,
  gdtable,
  sptree,
  sub_coverage,
  split,
  tree_id_tab,
  up_one_node = up_one_node,
  pdfwid = pdfwid,
  pdflen = pdflen
)
}
\arguments{
\item{x}{the row of phyto_node to be tested}

\item{phyto_node}{a table with the first column as the phyto id given by Tree2GD and 
the second column as the node id from \code{ape}.}

\item{trees}{a list of objects of class "\code{phylo}".
A list of gene trees to be examined.}

\item{gdtable}{a \code{data.frame} of the gd table file provided by Tree2GD.}

\item{sptree}{the species tree used to do tree reconciliation.}

\item{sub_coverage}{a numeric specify the species coverage level. 
If <= 1, treated as percentage. If > 1, treated as number of species.}

\item{split}{a character. The symbol used to separate species name and the 
sequence number in the gene family trees.}

\item{tree_id_tab}{a table for the tree id during Tree2GD and the name of the 
corresponding tree file. The first column is the id, and the second column
is the tree names.}

\item{up_one_node}{logical. Whether to choose tips of one node deeper as the 
closest outgroup, as a second step of investigation. Default \code{FALSE}.}

\item{pdfwid}{the width of pdf file for subtrees. Default \code{10}.}

\item{pdflen}{the length of pdf file for subtrees. Default \code{30}.}
}
\description{
Main internal function for \code{filter_GD_by_coverage} 
    to identify for each gd whether the two 
    duplicated subclades contains species number more than a given level.
}
\seealso{
\code{\link{filter_GD_by_coverage}}
}
\keyword{function}
\keyword{internal}
