% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sp_for_gd.R
\name{.sum_sp_gd}
\alias{.sum_sp_gd}
\title{Internal function}
\usage{
.sum_sp_gd(xx, phyto)
}
\arguments{
\item{xx}{a gd id}

\item{phyto}{a data.frame of gd table of a specific phyto node.}
}
\description{
Internal function of summarize_sp_for_GD
}
