% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{make.idTable.phyto}
\alias{make.idTable.phyto}
\title{generate node id table for the reference of phyto id and ape node id}
\usage{
make.idTable.phyto(tree)
}
\arguments{
\item{tree}{the tree for making id table.}
}
\description{
internal function
}
