% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.concor.R
\name{plot.concor}
\alias{plot.concor}
\title{Plot Concordances}
\usage{
\method{plot}{concor}(x, levelcol = NULL, thres = NULL,
  type = c("percent", "count"), node.cex = 0.5, node.adj = c(0.5, 0.5),
  node.frame = "rect", ...)
}
\arguments{
\item{x}{an object from "\code{concor.node}".}

\item{levelcol}{a character vector specifying the colors for node labels showing different
levels of concordances. Defaults as \code{NULL}.}

\item{thres}{a numeric vector specifying the levels for node labels to be plotted with
different colors. Must be consistent with \code{type}. Defaults as \code{NULL}.}

\item{type}{a character specifying the type of node labels. Can be either "percent" or "count".}

\item{node.cex}{a numeric specifying the scale of node labels. Defaults to \code{0.5}.}

\item{node.adj}{a numeric vector specifying the position of node labels. Defaults
to \code{c(0.5, 0.5)}. Also see \code{plot.phylo}.}

\item{node.frame}{a character specifying the type of frame of node labels. Must be one of
"rect", "circle", "none", or unambiguous abbreviation. Also see \code{plot.phylo}.
Defaults to "\code{rect}".}

\item{...}{other arguments to be passed to \code{plot} or to \code{plot.phylo}.}
}
\description{
This function plot the \code{ref} tree with the percents or counts of
concordances as node labels.
}
\details{
\code{thres} is required with \code{levelcol} arg. On the other hand, \code{levelcol}
    can left as default when \code{thres} is set, then \code{brewer.pal} (in \code{RColorBrewer}
    pcakage) will be called to produce \code{levelcol}. Length of \code{thres} must be one
    more than \code{levelcol}.
}
\seealso{
\code{\link{concor.node}}, \code{\link[ape]{plot.phylo}}
}
