% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFilteredGD.R
\name{plotFilteredGD}
\alias{plotFilteredGD}
\title{plot the gd result corrected by filter_GD_by_coverage
    and denominator}
\usage{
plotFilteredGD(
  tree,
  gd,
  deno,
  type = "percent",
  thres_node,
  thres_count = 100,
  thres_percent = 5,
  color = c("red", "yellow"),
  tip.offset = NULL,
  xlim = NULL
)
}
\arguments{
\item{tree}{the species tree used to do tree reconciliation.}

\item{gd}{the summary table of \code{filter_GD_by_coverage}.}

\item{deno}{the result of \code{denominator}.}

\item{type}{can be one of "percent", "count", or "both", determines 
the type of threshod value used for node color. Defaults to \code{"percent"}.}

\item{thres_node}{optional. A numeric specify which node 
(in ape id) are used as threshold to give colors on node labels.}

\item{thres_count}{nodes with counts above this threshold will be plotted 
as red. Defaults to \code{100}}

\item{thres_percent}{nodes with percent above this threshold will be plotted 
as red. Defaults to \code{5}}

\item{color}{two characters specify the colors for node labels. The first is 
for node above the threshold, and the second is for those below.}

\item{tip.offset}{offset of tip labels in \code{ggtree}. Defaults to \code{NULL}.}

\item{xlim}{used as the x scale in \code{ggtree}.}
}
\description{
\code{plotFilteredGD} plot a tree with node labels as
    the corrected values (both count and percent). Only nodes both 
    shown in both \code{gd} and \code{deno} have node labels. Only 
    nodes with deno > 0 are shown.
}
