% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.tree.R
\name{rewrite.tree}
\alias{rewrite.tree}
\title{rewrite.tree}
\usage{
rewrite.tree(tree)
}
\arguments{
\item{tree}{tree}
}
\description{
Internal function of \code{youshu}. Make the numbering of nodes and tips of a reordered tree
    as re-read in.
}
\seealso{
\code{\link{root_dist}}, \code{\link[ape]{root.phylo}}
}
\keyword{internal}
