% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{root.dist}
\alias{root.dist}
\title{internal function}
\usage{
root.dist(z, tar, bp = FALSE)
}
\arguments{
\item{z}{an object of class "\code{phylo}"}

\item{tar}{target tips in the current analysis}

\item{bp}{a logical indicating whether to retain
the original bp values in the input tree.
Currently desinged for backboneBP.}
}
\description{
Re-numbering the tips and nodes of ape tree objects.
    Internal function of \code{rePhylo}.
}
\seealso{
\code{\link{cladeFilter}}, \code{\link{concor.node}}, \code{\link{backboneBP}}, \code{\link{concor.trees}}
}
\keyword{function}
\keyword{internal}
