% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{select.backbone}
\alias{select.backbone}
\title{internal function}
\usage{
select.backbone(tre, mrcalist)
}
\arguments{
\item{tre}{a tree}

\item{mrcalist}{mrcalist}
}
\description{
Identify backbone nodes based on \code{taxa} and
    make other bp values as NA. Internal function of \code{rePhylo}.
}
\seealso{
\code{\link{backboneBP}}
}
\keyword{internal}
