% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sp_for_gd.R
\name{summarize_sp_for_GD}
\alias{summarize_sp_for_GD}
\title{Summarize number of species supporting gd}
\usage{
summarize_sp_for_GD(phyto, pdf = TRUE, pdfwid = 5, pdflen = NULL)
}
\arguments{
\item{phyto}{a data.frame of gd table of a specific phyto node.}

\item{pdf}{a logical specifying whether to export plots into pdf files.
Defaults to \code{TRUE}.}

\item{pdfwid}{a numeric for width of pdf file. Defaults to 5.}

\item{pdflen}{optional. Length of pdf file. Defaults to \code{NULL}.}
}
\description{
Summarize the number of gd supported by each species, 
    and the number of gd supported by how much species.
}
