% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMulticf.R
\name{summarizeMulticf}
\alias{summarizeMulticf}
\title{Summarize \code{cladeFilter} Results with Multiple Groupings}
\usage{
summarizeMulticf(
  x,
  trees,
  plot.tree = TRUE,
  write.table = TRUE,
  rm.col = c(rmOnce = "red", rmMulti = "blue"),
  ...
)
}
\arguments{
\item{x}{an object from "\code{cladeFilter}".}

\item{trees}{a list of objects of class "\code{phylo}".
A list of test trees such as single gene trees.}

\item{plot.tree}{a logical specifying whether to plot \code{trees} with coloring tips
to show tips suggested to be removed by \code{cladeFilter} for all groupings.
Defaults to \code{TRUE}.}

\item{write.table}{a logical specifying whether to write out text files
containing the summarized number and names of sequences suggested to be removed
by \code{cladeFilter} for all groupings. Defaults to \code{TRUE}.}

\item{rm.col}{a character vector specifying the colors for tips removing by one grouping
(defalts to \code{red}) or by multiple groupings (defaults to \code{blue}).}

\item{...}{other arguments to be passed to \code{plot} or to \code{plot.phylo}.}
}
\description{
This function summarizes the scenarios of all the groupings for each of
the \code{trees}, and generates barplots showing the numbers of removing sequences
by species (datasets), genetrees (genes) and groupings into pdf files.
}
\examples{
\dontrun{
data(Brassidata)
trees <- Brassidata$trees
taxa <- Brassidata$taxaTable
# perform cladeFilter analysis
ures <- cladeFilter(trees = trees, taxa = taxa, level = 2)
# summarize cladeFilter results with multiple groupings
usum <- summarizeMulticf(ures, trees, plot.tree = FALSE, write.table = FALSE)
# to make barplots
utest <- barplotMulticf(x = usum, reorder="up", plot = TRUE, pdf=FALSE)
}
}
\seealso{
\code{\link{cladeFilter}}, \code{\link{barplotMulticf}}
}
